refactored browsetags component into tagslist
